+++ rabbithole/src/Render/Vulkan/VulkanBuffer.h
@@ -7,7 +7,7 @@ struct VulkanBufferInfo
 {
 	BufferUsageFlags usageFlags;
 	MemoryAccess	 memoryAccess;
-	uint32_t		 size;
+	uint64_t		 size;
 };
 
 class VulkanBuffer : public AllocatedResource
@@ -25,15 +25,15 @@ public:
 	void* Map();
 	void  Unmap();
 	void  FillBuffer(void* data);
-	void  FillBuffer(void* data, size_t size);
-	void  FillBuffer(void* data, size_t offset, size_t size);
+	void  FillBuffer(void* data, uint64_t size);
+	void  FillBuffer(void* data, uint64_t offset, uint64_t size);
 
 public:
 	inline const VulkanBufferInfo	GetInfo() const { return m_Info; }
 	inline void*					GetHostVisibleData() { return m_HostVisibleData; }
 	inline VkBuffer					GetBuffer() { return m_Buffer; }
 	uint32_t						GetID() { return m_Id; }
-	inline size_t					GetSize() { return m_Size; }
+	inline uint64_t					GetSize() { return m_Size; }
 
 private:
 	void CreateBufferResource();
@@ -46,5 +46,5 @@ private:
 	void*					m_HostVisibleData = nullptr;
 
 	std::string				m_Name;
-	size_t					m_Size;
+	uint64_t				m_Size;
 };
